Format changes

+++ DBModel/Models/ProfitLossReport.cs
@@ -11,23 +11,23 @@ namespace DBModel.Models
         int brojKutija;
         int prosecanTrosakPoBeracu;
         int trosak;
-        int netoOtkup;
+        decimal netoOtkup;
         int prosecnaPC;
         int prosecnaCenaKost;
         int prihod;
         int profit;
-        decimal trosakProc;
+        int trosakProc;
 
         public DateTime Datum { get => datum; set => datum = value; }
         public int BrojBeraca { get => brojBeraca; set => brojBeraca = value; }
         public int BrojKutija { get => brojKutija; set => brojKutija = value; }
         public int ProsecanTrosakPoBeracu { get => prosecanTrosakPoBeracu; set => prosecanTrosakPoBeracu = value; }
         public int Trosak { get => trosak; set => trosak = value; }
-        public int NetoOtkup { get => netoOtkup; set => netoOtkup = value; }
+        public decimal NetoOtkup { get => netoOtkup; set => netoOtkup = value; }
         public int ProsecnaPC { get => prosecnaPC; set => prosecnaPC = value; }
         public int ProsecnaCenaKost { get => prosecnaCenaKost; set => prosecnaCenaKost = value; }
         public int Prihod { get => prihod; set => prihod = value; }
         public int Profit { get => profit; set => profit = value; }
-        public decimal TrosakProc { get => trosakProc; set => trosakProc = value; }
+        public int TrosakProc { get => trosakProc; set => trosakProc = value; }
     }
 }
